Python click integration
