+++ Dockerfile
@@ -1,35 +1,33 @@
 FROM python:3.7.3
 
 RUN echo '---> Setting up build environment' \
-    && apt update -y \
-	&& apt install -y git gcc make libssl-dev libffi-dev libsqlite3-dev libpq-dev locales
+    && apt-get update -y \
+	&& apt-get install -y git gcc make libssl-dev libffi-dev libsqlite3-dev libpq-dev locales
 
-
-RUN echo ' ---> Setting up user environment' \
+RUN echo '---> Setting up user environment' \
     && adduser -u 1000 pili \
     && mkdir /app \
-    && chown pili /app
+    && chown -R pili:pili /app
+
+COPY . /app
+WORKDIR /app
 
 RUN localedef -c -f UTF-8 -i en_US en_US.UTF-8
 ENV LANG=en_US.UTF-8
+ENV TERM=xterm-256color
 
-COPY requirements.txt /app/requirements.txt
-
-RUN echo ' ---> Installing package dependencies' \
+RUN echo '---> Installing package dependencies' \
 	&& pip install --upgrade pip \
-	&& pip install -r /app/requirements.txt
+	&& pip install -r /app/requirements.txt \
+	&& pip install -e .
 
-COPY . /app
-
-WORKDIR /app
-
-RUN echo ' ---> Clean up build environment' \
-	&& apt autoremove -y \
-	&& apt clean -y \
+RUN echo '---> Clean up build environment' \
+	&& apt-get autoremove -y \
+	&& apt-get clean -y \
     && rm -rf ~/.cache \
     && rm -rf /var/cache/apt \
     && sh -c 'find . | grep -E "(_pycache_|\.pyc|\.pyo$)" | xargs rm -rf'
 
 USER pili
 
-CMD ["python", "manage.py", "runserver", "--host=0.0.0.0"]
+CMD ["pili", "uwsgi", "--config=production", "--section=production"]
