+++ etc/uwsgi.ini
@@ -1,4 +1,8 @@
-[uwsgi]
+#
+# Main section
+#
+
+[default]
 # upstream
 http-socket = 0.0.0.0:8080
 
@@ -9,7 +13,7 @@ master = true
 listen = 10
 
 # app entry point
-module = manage:app_dispatch
+module = pili.entrypoints.wsgi:application
 
 # application's absolute path
 app_dir = /app
@@ -55,23 +59,51 @@ memory-report = true
 # clear environment on exit
 #vacuum = true
 
-[development]
-ini = :uwsgi
+#
+# Sections by workers
+#
 
+[solo]
+ini = :default
+workers = 1
+threads = 1
+
+[duet]
+ini = :default
 workers = 2
 threads = 1
 
-
-[production]
-ini = :development
-
+[quartet]
+ini = :default
 workers = 4
 threads = 2
 
+[octet]
+ini = :default
+workers = 8
+threads = 4
 
-[unix]
-ini = :production
+[choir]
+ini = :default
+workers = 16
+threads = 8
 
+[orchestra]
+ini = :default
+workers = 32
+threads = 16
 
-[heroku]
-ini = :production
+#
+# Sections by environments
+#
+
+[testing]
+ini = :solo
+python-autoreload = true
+
+[development]
+ini = :solo
+python-autoreload = true
+
+[production]
+ini = :quartet
