--- flower-entrypoint.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-
-flower -A celery_worker.celery --broker_api=${FLOWER_BROKER_API}
