+++ pili/auth/views.py
@@ -1,13 +1,9 @@
 from flask import current_app, flash, redirect, render_template, request, url_for
 from flask_login import current_user, login_required, login_user, logout_user
 
-from .. import db
-from ..decorators import permission_required
-from ..email import send_email
-from ..filters import generate_password
-from ..models import Permission, Post, Role, User
-from . import auth
-from .forms import (
+from pili.app import db
+from pili.auth import auth
+from pili.auth.forms import (
     ChangeEmailForm,
     ChangePasswordForm,
     InviteAcceptForm,
@@ -17,6 +13,10 @@ from .forms import (
     PasswordResetRequestForm,
     RegistrationForm,
 )
+from pili.decorators import permission_required
+from pili.email import send_email
+from pili.filters import generate_password
+from pili.models import Permission, Post, Role, User
 
 
 @auth.before_app_request
