+++ pili/ctrl/__init__.py
@@ -2,8 +2,8 @@ from flask import Blueprint
 
 ctrl = Blueprint('ctrl', __name__)
 
-from . import views, errors
-from ..models import Permission
+from pili.ctrl import views, errors
+from pili.models import Permission
 
 
 @ctrl.app_context_processor
