+++ pili/main/__init__.py
@@ -2,8 +2,8 @@ from flask import Blueprint
 
 main = Blueprint('main', __name__)
 
-from . import views, errors
-from ..models import Permission
+from pili.main import views, errors
+from pili.models import Permission
 
 
 @main.app_context_processor
