+++ setup.py
@@ -16,4 +16,7 @@ setup(
     include_package_data=True,
     packages=find_packages(exclude=['tests']),
     namespace_packages=['pili'],
+    entry_points={
+          'console_scripts': ['pili = pili.entrypoints.commands:cli']
+    },
 )
